LayoutTests/fast/spatial-navigation/snav-only-clipped-overflow-content-expected.txt

That is is a normal link.

This is a scrollable Div created with the CSS property overflow.









... and here we have a clipped overflow link.

And this is another normal link_2.

PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS true is true
